fix: Update VentasByVendedor component to correctly set ventas data
The VentasByVendedor component was updated to correctly set the ventas data received from the API response. Previously, only the response data was being set, but now the "data" property of the response is being assigned to the ventas state variable. This ensures that the ventas data is properly updated and displayed in the component.